Rename create_aws_resources to manage_aws_resources
