+++ delete_redshift_cluster.py
@@ -5,7 +5,7 @@ Delete Redshift Cluster and Data Warehouse IAM role
 """
 import dotenv
 import boto3
-from utils.create_aws_resources import (
+from utils.manage_aws_resources import (
     delete_redshift_cluster,
     delete_role)
 from config import Config
@@ -29,7 +29,7 @@ def main():
     role_response = delete_role(iam_client=iam, role_name=config.iam_role_name)
     # Redshift cluster
     redshift = boto3.client('redshift')
-    cluster_props = delete_redshift_cluster(
+    _ = delete_redshift_cluster(
         redshift_client=redshift,
         cluster_identifier=config.cluster_identifier)
     # Remove db_host from CONFIG_FILE
